Moved Popup and TitleBarUtility to utility package
