+++ src/main/java/com/gearshiftgaming/se_mod_manager/frontend/view/ModProfileManagerView.java
@@ -4,7 +4,8 @@ import com.gearshiftgaming.se_mod_manager.backend.models.ModProfile;
 import com.gearshiftgaming.se_mod_manager.backend.models.utility.MessageType;
 import com.gearshiftgaming.se_mod_manager.frontend.domain.UiService;
 import com.gearshiftgaming.se_mod_manager.frontend.models.ModProfileCell;
-import com.gearshiftgaming.se_mod_manager.frontend.view.helper.TitleBarUtility;
+import com.gearshiftgaming.se_mod_manager.frontend.view.utility.TitleBarUtility;
+import com.gearshiftgaming.se_mod_manager.frontend.view.utility.Popup;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.Parent;
