+++ src/main/java/com/gearshiftgaming/se_mod_manager/frontend/view/utility/Popup.java
@@ -1,8 +1,7 @@
-package com.gearshiftgaming.se_mod_manager.frontend.view;
+package com.gearshiftgaming.se_mod_manager.frontend.view.utility;
 
 import com.gearshiftgaming.se_mod_manager.backend.models.utility.MessageType;
 import com.gearshiftgaming.se_mod_manager.backend.models.utility.Result;
-import com.gearshiftgaming.se_mod_manager.frontend.view.helper.TitleBarUtility;
 import javafx.beans.value.ChangeListener;
 import javafx.event.ActionEvent;
 import javafx.geometry.Insets;
