+++ src/main/resources/view/mod-table-context-bar.fxml
@@ -15,8 +15,6 @@
 <?import javafx.scene.shape.Rectangle?>
 <?import javafx.scene.text.Font?>
 <?import org.kordamp.ikonli.javafx.FontIcon?>
-<?import org.kordamp.ikonli.carbonicons.*?>
-<?import org.kordamp.ikonli.coreui.*?>
 
 <VBox fx:id="menuBarRoot" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1">
      <children>
@@ -87,7 +85,7 @@
                          <Label minWidth="-Infinity" text="Active Mod Count" />
                          <StackPane alignment="CENTER_LEFT" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="28.0" prefWidth="63.0">
                              <children>
-                                 <Rectangle arcHeight="5.0" arcWidth="5.0" fill="#1f93ff00" height="28.0" stroke="BLACK" strokeType="INSIDE" width="63.0" />
+                                 <Rectangle fx:id="activeModCountRectangle" arcHeight="5.0" arcWidth="5.0" fill="#1f93ff00" height="28.0" stroke="BLACK" strokeType="INSIDE" width="63.0" />
                         <Label fx:id="activeModCount" text="0" StackPane.alignment="CENTER_RIGHT">
                            <font>
                               <Font size="14.0" />
@@ -111,7 +109,7 @@
                          <Label minWidth="-Infinity" text="Mod Conflicts" />
                          <StackPane alignment="CENTER_LEFT" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="28.0" prefWidth="63.0">
                              <children>
-                                 <Rectangle arcHeight="5.0" arcWidth="5.0" fill="#1f93ff00" height="28.0" stroke="BLACK" strokeType="INSIDE" width="63.0" />
+                                 <Rectangle fx:id="modConflictRectangle" arcHeight="5.0" arcWidth="5.0" fill="#1f93ff00" height="28.0" stroke="BLACK" strokeType="INSIDE" width="63.0" />
                         <Label fx:id="modConflicts" text="0" StackPane.alignment="CENTER_RIGHT">
                            <padding>
                               <Insets right="10.0" />
